requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14.0
djoser>=2.2.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
gunicorn>=21.2.0
